feat: Moving methods from CIUtils to StepUtil
