+++ src/modules/75-ci/components/PipelineSteps/CIStep/ArtifactStepCommon.tsx
@@ -9,7 +9,7 @@ import React from 'react'
 import { get } from 'lodash-es'
 import { getMultiTypeFromValue, MultiTypeInputType } from '@wings-software/uicore'
 import type { ConnectorInfoDTO } from 'services/cd-ng'
-import { shouldRenderRunTimeInputView } from '@pipeline/utils/CIUtils'
+import { shouldRenderRunTimeInputView } from '@pipeline/components/PipelineStudio/StepUtil'
 import { Connectors } from '@connectors/constants'
 import type { ECRStepProps } from '../ECRStep/ECRStep'
 import type { GCRStepProps } from '../GCRStep/GCRStep'
