+++ src/modules/75-ci/components/PipelineSteps/CIStep/ConnectorRefWithImage.tsx
@@ -16,7 +16,8 @@ import { FormMultiTypeConnectorField } from '@connectors/components/ConnectorRef
 import { useStrings } from 'framework/strings'
 import { MultiTypeTextField } from '@common/components/MultiTypeText/MultiTypeText'
 import { useVariablesExpression } from '@pipeline/components/PipelineStudio/PiplineHooks/useVariablesExpression'
-import { useGitScope, shouldRenderRunTimeInputViewWithAllowedValues } from '@pipeline/utils/CIUtils'
+import { shouldRenderRunTimeInputViewWithAllowedValues } from '@pipeline/components/PipelineStudio/StepUtil'
+import { useGitScope } from '@pipeline/utils/CIUtils'
 import { StepViewType } from '@pipeline/components/AbstractSteps/Step'
 import { renderMultiTypeInputWithAllowedValues, getOptionalSubLabel } from './CIStepOptionalConfig'
 import { AllMultiTypeInputTypesForStep } from './StepUtils'
