+++ src/modules/75-ci/components/PipelineSteps/Dependency/DependencyInputSet.tsx
@@ -10,7 +10,7 @@ import { connect } from 'formik'
 import { getMultiTypeFromValue, MultiTypeInputType, FormikForm } from '@wings-software/uicore'
 import { useStrings } from 'framework/strings'
 import { useVariablesExpression } from '@pipeline/components/PipelineStudio/PiplineHooks/useVariablesExpression'
-import { shouldRenderRunTimeInputView } from '@pipeline/utils/CIUtils'
+import { shouldRenderRunTimeInputView } from '@pipeline/components/PipelineStudio/StepUtil'
 import StepCommonFieldsInputSet from '@ci/components/PipelineSteps/StepCommonFields/StepCommonFieldsInputSet'
 import { Connectors } from '@connectors/constants'
 import type { DependencyProps } from './Dependency'
